Some Fixes #1
