+++ one_time_setup.bat
@@ -1,0 +1,2 @@
+@echo off
+cmd /c "cd /d .\virajs_excel & py one_time_setup.py"
