+++ run_algo.bat
@@ -1,2 +1,4 @@
-@echo on
-cmd /k "cd /d C:\ecomsense\env\Scripts & .\activate & cd /d C:\ecomsense\env\virajs-excel\virajs_excel & py excel_magic.py"
+@echo off
+echo Starting Program...
+cmd /c "..\venv\Scripts\activate & cd /d .\virajs_excel & py main.py"
+
