--- virajs_excel/one_time_setup.bat
@@ -1,2 +1,0 @@
-@echo on
-cmd /k "cd /d C:\ecomsense\env\virajs-excel\virajs_excel & py one_time_setup.py"
