+++ virajs_excel/test/example_orders.json
@@ -1,0 +1,195 @@
+{
+    "status": "success",
+    "data": {
+        "net": [
+            {
+                "tradingsymbol": "LEADMINI17DECFUT",
+                "exchange": "MCX",
+                "instrument_token": 53496327,
+                "product": "NRML",
+                "quantity": 1,
+                "overnight_quantity": 0,
+                "multiplier": 1000,
+                "average_price": 161.05,
+                "close_price": 0,
+                "last_price": 161.05,
+                "value": -161050,
+                "pnl": 0,
+                "m2m": 0,
+                "unrealised": 0,
+                "realised": 0,
+                "buy_quantity": 1,
+                "buy_price": 161.05,
+                "buy_value": 161050,
+                "buy_m2m": 161050,
+                "sell_quantity": 0,
+                "sell_price": 0,
+                "sell_value": 0,
+                "sell_m2m": 0,
+                "day_buy_quantity": 1,
+                "day_buy_price": 161.05,
+                "day_buy_value": 161050,
+                "day_sell_quantity": 0,
+                "day_sell_price": 0,
+                "day_sell_value": 0
+            },
+            {
+                "tradingsymbol": "GOLDGUINEA17DECFUT",
+                "exchange": "MCX",
+                "instrument_token": 53505799,
+                "product": "NRML",
+                "quantity": 0,
+                "overnight_quantity": 3,
+                "multiplier": 1,
+                "average_price": 0,
+                "close_price": 23232,
+                "last_price": 23355,
+                "value": 801,
+                "pnl": 801,
+                "m2m": 276,
+                "unrealised": 801,
+                "realised": 0,
+                "buy_quantity": 4,
+                "buy_price": 23139.75,
+                "buy_value": 92559,
+                "buy_m2m": 93084,
+                "sell_quantity": 4,
+                "sell_price": 23340,
+                "sell_value": 93360,
+                "sell_m2m": 93360,
+                "day_buy_quantity": 1,
+                "day_buy_price": 23388,
+                "day_buy_value": 23388,
+                "day_sell_quantity": 4,
+                "day_sell_price": 23340,
+                "day_sell_value": 93360
+            },
+            {
+                "tradingsymbol": "SBIN",
+                "exchange": "NSE",
+                "instrument_token": 779521,
+                "product": "CO",
+                "quantity": 0,
+                "overnight_quantity": 0,
+                "multiplier": 1,
+                "average_price": 0,
+                "close_price": 0,
+                "last_price": 308.4,
+                "value": -2,
+                "pnl": -2,
+                "m2m": -2,
+                "unrealised": -2,
+                "realised": 0,
+                "buy_quantity": 1,
+                "buy_price": 311,
+                "buy_value": 311,
+                "buy_m2m": 311,
+                "sell_quantity": 1,
+                "sell_price": 309,
+                "sell_value": 309,
+                "sell_m2m": 309,
+                "day_buy_quantity": 1,
+                "day_buy_price": 311,
+                "day_buy_value": 311,
+                "day_sell_quantity": 1,
+                "day_sell_price": 309,
+                "day_sell_value": 309
+            }
+        ],
+        "day": [
+            {
+                "tradingsymbol": "GOLDGUINEA17DECFUT",
+                "exchange": "MCX",
+                "instrument_token": 53505799,
+                "product": "NRML",
+                "quantity": -3,
+                "overnight_quantity": 0,
+                "multiplier": 1,
+                "average_price": 23340,
+                "close_price": 23232,
+                "last_price": 23355,
+                "value": 69972,
+                "pnl": -93,
+                "m2m": -93,
+                "unrealised": -93,
+                "realised": 0,
+                "buy_quantity": 1,
+                "buy_price": 23388,
+                "buy_value": 23388,
+                "buy_m2m": 23388,
+                "sell_quantity": 4,
+                "sell_price": 23340,
+                "sell_value": 93360,
+                "sell_m2m": 93360,
+                "day_buy_quantity": 1,
+                "day_buy_price": 23388,
+                "day_buy_value": 23388,
+                "day_sell_quantity": 4,
+                "day_sell_price": 23340,
+                "day_sell_value": 93360
+            },
+            {
+                "tradingsymbol": "LEADMINI17DECFUT",
+                "exchange": "MCX",
+                "instrument_token": 53496327,
+                "product": "NRML",
+                "quantity": 1,
+                "overnight_quantity": 0,
+                "multiplier": 1000,
+                "average_price": 161.05,
+                "close_price": 0,
+                "last_price": 161.05,
+                "value": -161050,
+                "pnl": 0,
+                "m2m": 0,
+                "unrealised": 0,
+                "realised": 0,
+                "buy_quantity": 1,
+                "buy_price": 161.05,
+                "buy_value": 161050,
+                "buy_m2m": 161050,
+                "sell_quantity": 0,
+                "sell_price": 0,
+                "sell_value": 0,
+                "sell_m2m": 0,
+                "day_buy_quantity": 1,
+                "day_buy_price": 161.05,
+                "day_buy_value": 161050,
+                "day_sell_quantity": 0,
+                "day_sell_price": 0,
+                "day_sell_value": 0
+            },
+            {
+                "tradingsymbol": "SBIN",
+                "exchange": "NSE",
+                "instrument_token": 779521,
+                "product": "CO",
+                "quantity": 0,
+                "overnight_quantity": 0,
+                "multiplier": 1,
+                "average_price": 0,
+                "close_price": 0,
+                "last_price": 308.4,
+                "value": -2,
+                "pnl": -2,
+                "m2m": -2,
+                "unrealised": -2,
+                "realised": 0,
+                "buy_quantity": 1,
+                "buy_price": 311,
+                "buy_value": 311,
+                "buy_m2m": 311,
+                "sell_quantity": 1,
+                "sell_price": 309,
+                "sell_value": 309,
+                "sell_m2m": 309,
+                "day_buy_quantity": 1,
+                "day_buy_price": 311,
+                "day_buy_value": 311,
+                "day_sell_quantity": 1,
+                "day_sell_price": 309,
+                "day_sell_value": 309
+            }
+        ]
+    }
+}
